Split image workspace foundations
