+++ frontend/app/(core)/(workspace)/app/image/ImageWorkspace.tsx
@@ -8,30 +8,20 @@ import dynamic from 'next/dynamic';
 import { useCallback, useEffect, useMemo, useState, useRef } from 'react';
 import { usePathname, useSearchParams } from 'next/navigation';
 import { Plus } from 'lucide-react';
-import type {
-  ChangeEvent,
-  FormEvent,
-  MouseEvent as ReactMouseEvent,
-} from 'react';
-import type { PricingSnapshot } from '@maxvideoai/pricing';
+import type { FormEvent } from 'react';
 import { GalleryRail } from '@/components/GalleryRail';
 import { Button, ButtonLink } from '@/components/ui/Button';
 import { EngineSelect } from '@/components/ui/EngineSelect';
 import { Composer, type AssetFieldConfig, type ComposerAttachment } from '@/components/Composer';
 import { ImageSettingsBar } from '@/components/ImageSettingsBar';
 import { ImageAdvancedSettings } from '@/components/ImageAdvancedSettings';
-import type { AssetLibraryBrowserProps, AssetLibrarySource } from '@/components/library/AssetLibraryBrowser';
 import { runImageGeneration, useInfiniteJobs, saveImageToLibrary } from '@/lib/api';
 import { suggestDownloadFilename, triggerAppDownload } from '@/lib/download';
-import { translateError } from '@/lib/error-messages';
 import type {
   CharacterReferenceSelection,
-  CharacterReferencesResponse,
   ImageGenerationMode,
-  GeneratedImage,
 } from '@/types/image-generation';
-import type { EngineCaps } from '@/types/engines';
-import type { GroupSummary, GroupMemberSummary } from '@/types/groups';
+import type { GroupSummary } from '@/types/groups';
 import type { Job } from '@/types/jobs';
 import type { VideoGroup } from '@/types/video-groups';
 import type { MediaLightboxEntry } from '@/components/MediaLightbox';
@@ -39,7 +29,6 @@ import { ImageCompositePreviewDock, type ImageCompositePreviewEntry } from '@/co
 import { buildVideoGroupFromImageRun } from '@/lib/image-groups';
 import { useI18n } from '@/lib/i18n/I18nProvider';
 import { prepareImageFileForUpload } from '@/lib/client-image-upload';
-import { normalizeUiLocale } from '@/lib/ltx-localization';
 import { formatAspectRatioLabel } from '@/lib/image/aspectRatios';
 import { FEATURES } from '@/content/feature-flags';
 import {
@@ -50,7 +39,6 @@ import {
   isSupportedImageMime,
 } from '@/lib/image/formats';
 import {
-  MAX_REFERENCE_IMAGES,
   clampRequestedImageCount,
   getAspectRatioOptions,
   getDefaultAspectRatio,
@@ -74,872 +62,62 @@ import { readLastKnownUserId } from '@/lib/last-known';
 import { readBrowserSession } from '@/lib/supabase-auth-cleanup';
 import { hasSupabaseAuthCookie } from '@/lib/supabase-session-hint';
 import { normalizeJobSurface } from '@/lib/job-surface-normalize';
-import { isPlaceholderMediaUrl, resolvePreferredMediaUrl } from '@/lib/media';
 import { groupJobsIntoSummaries } from '@/lib/job-groups';
 import { countResolvedVisualSlots } from '@/lib/group-progress';
+import { ImageLibraryModal } from './_components/ImageLibraryModal';
+import { useImageWorkspaceDesktopLayout } from './_hooks/useImageWorkspaceDesktopLayout';
+import {
+  createCharacterReferenceSlot,
+  mergeCharacterReferencesIntoSlots,
+  parseCharacterReferenceEntry,
+} from './_lib/image-workspace-character-references';
+import {
+  DEFAULT_COPY,
+  formatTemplate,
+  mergeCopy,
+  type ImageWorkspaceCopy,
+} from './_lib/image-workspace-copy';
+import {
+  buildCompletedGroup,
+  buildPendingGroup,
+  mapJobToHistoryEntry,
+} from './_lib/image-workspace-history';
+import { parsePersistedImageComposerState } from './_lib/image-workspace-persistence';
+import {
+  buildCustomImageSize,
+  findImageEngine,
+  formatImageSizeLabel,
+  formatQualityLabel,
+} from './_lib/image-workspace-utils';
+import {
+  DEFAULT_UPLOAD_LIMIT_MB,
+  DEFAULT_VISIBLE_REFERENCE_SLOTS,
+  IMAGE_COMPOSER_STORAGE_DEBOUNCE_MS,
+  IMAGE_COMPOSER_STORAGE_KEY,
+  IMAGE_COMPOSER_STORAGE_VERSION,
+  MAX_REFERENCE_SLOTS,
+  QUICK_IMAGE_COUNT_OPTIONS,
+  type AssetsResponse,
+  type HistoryEntry,
+  type ImageEngineOption,
+  type ImageLibraryModalState,
+  type LibraryAsset,
+  type PersistedImageComposerState,
+  type PersistedReferenceSlot,
+  type ReferenceSlotValue,
+  type PricingEstimateResponse,
+  type UploadFailure,
+} from './_lib/image-workspace-types';
 
-const AssetLibraryBrowser = dynamic<AssetLibraryBrowserProps>(
-  () => import('@/components/library/AssetLibraryBrowser').then((mod) => mod.AssetLibraryBrowser),
-  { ssr: false }
-);
+export type { ImageEngineOption } from './_lib/image-workspace-types';
 
 const GroupViewerModal = dynamic(
   () => import('@/components/groups/GroupViewerModal').then((mod) => mod.GroupViewerModal),
   { ssr: false }
 );
 
-interface ImageWorkspaceCopy {
-  hero: {
-    eyebrow: string;
-    title: string;
-    subtitle: string;
-  };
-  preview: {
-    eyebrow: string;
-    title: string;
-    emptyTitle: string;
-    emptyDescription: string;
-    download: string;
-    copy: string;
-    copied: string;
-    cta: string;
-  };
-  engine: {
-    eyebrow: string;
-    priceCalculating: string;
-    priceLabel: string;
-  };
-  modeTabs: {
-    generate: string;
-    edit: string;
-  };
-  composer: {
-    promptLabel: string;
-    promptPlaceholder: string;
-    promptPlaceholderWithImage: string;
-    charCount: string;
-    presetsHint: string;
-    numImagesLabel: string;
-    numImagesCount: string;
-    numImagesUnit: {
-      singular: string;
-      plural: string;
-    };
-    aspectRatioLabel: string;
-    aspectRatioHint: string;
-    aspectRatioAutoNote: string;
-    resolutionLabel: string;
-    resolutionHint: string;
-    resolutionLockedLabel: string;
-    seedLabel: string;
-    seedPlaceholder: string;
-    outputFormatLabel: string;
-    outputFormatHint: string;
-    qualityLabel: string;
-    qualityHint: string;
-    customWidthLabel: string;
-    customHeightLabel: string;
-    maskUrlLabel: string;
-    maskUrlPlaceholder: string;
-    enableWebSearchLabel: string;
-    enableWebSearchHint: string;
-    thinkingLevelLabel: string;
-    thinkingLevelHint: string;
-    limitGenerationsLabel: string;
-    limitGenerationsHint: string;
-    toggleEnabled: string;
-    toggleDisabled: string;
-    estimatedCost: string;
-    referenceLabel: string;
-    referenceHelper: string;
-    referenceNote: string;
-    referenceButton: string;
-    characterButton: string;
-    characterNote: string;
-    characterHiddenNotice: string;
-    characterLimitNotice: string;
-    referenceExpand: string;
-    referenceCollapse: string;
-    referenceSlotLabel: string;
-    referenceSlotHint: string;
-    referenceSlotNameFallback: string;
-    referenceSlotActions: {
-      upload: string;
-      library: string;
-      replace: string;
-      remove: string;
-    };
-  };
-  history: {
-    eyebrow: string;
-    empty: string;
-    runsLabel: {
-      zero: string;
-      other: string;
-    };
-    noPreview: string;
-    loadMore: string;
-    refreshing: string;
-  };
-  runButton: {
-    idle: string;
-    running: string;
-  };
-  errors: {
-    onlyImages: string;
-    unsupportedFormat: string;
-    uploadFailed: string;
-    fileTooLarge: string;
-    unauthorized: string;
-    promptMissing: string;
-    referenceMissing: string;
-    generic: string;
-  };
-  messages: {
-    success: string;
-    generatingInProgress: string;
-    savedToLibrary: string;
-    removedFromLibrary: string;
-  };
-  general: {
-    uploading: string;
-    cancelUpload: string;
-    emptyEngines: string;
-  };
-  library: {
-    button: string;
-    empty: string;
-    overlay: string;
-    unsupported: string;
-    supportedFormats: string;
-    assetFallback: string;
-    tabs: {
-      all: string;
-      upload: string;
-      generated: string;
-      character: string;
-      angle: string;
-      upscale: string;
-    };
-    modal: {
-      title: string;
-      description: string;
-      close: string;
-      error: string;
-      empty: string;
-      emptyCompatible: string;
-      emptyUploads: string;
-      emptyGenerated: string;
-      emptyCharacter: string;
-      emptyAngle: string;
-      emptyUpscale: string;
-    };
-  };
-  characterPicker: {
-    title: string;
-    description: string;
-    selected: string;
-    select: string;
-    done: string;
-    empty: string;
-    limitLabel: string;
-  };
-  authGate: {
-    title: string;
-    body: string;
-    primary: string;
-    secondary: string;
-    close: string;
-  };
-}
-
-const DEFAULT_COPY: ImageWorkspaceCopy = {
-  hero: {
-    eyebrow: 'Image workspace',
-    title: 'Generate Images',
-    subtitle: 'Create and edit high-fidelity images.',
-  },
-  preview: {
-    eyebrow: 'Generate preview',
-    title: 'Latest output',
-    emptyTitle: 'No preview yet.',
-    emptyDescription: 'Run a generation to see the latest image here.',
-    download: 'Download',
-    copy: 'Copy link',
-    copied: 'Copied',
-    cta: 'Generate image',
-  },
-  engine: {
-    eyebrow: 'Engine',
-    priceCalculating: 'Calculating price…',
-    priceLabel: '{amount} / run',
-  },
-  modeTabs: {
-    generate: 'Create',
-    edit: 'Edit',
-  },
-  composer: {
-    promptLabel: 'Prompt',
-    promptPlaceholder: 'Describe the image you’d like to generate...',
-    promptPlaceholderWithImage: 'Describe how the image should be edited / transformed…',
-    charCount: '{count} chars',
-    presetsHint: 'Tap a preset or start from scratch.',
-    numImagesLabel: 'Number of images',
-    numImagesCount: '{count} {unit}',
-    numImagesUnit: {
-      singular: 'image',
-      plural: 'images',
-    },
-    aspectRatioLabel: 'Aspect ratio',
-    aspectRatioHint: 'Choose a preset to match your frame.',
-    aspectRatioAutoNote: 'Auto lets the model decide the final crop.',
-    resolutionLabel: 'Resolution',
-    resolutionHint: 'Adjust output size here. Pricing updates automatically.',
-    resolutionLockedLabel: 'Resolution locked by this engine.',
-    seedLabel: 'Seed',
-    seedPlaceholder: 'Optional',
-    outputFormatLabel: 'Output format',
-    outputFormatHint: 'Choose the file type for the final images.',
-    qualityLabel: 'Quality',
-    qualityHint: 'Controls model fidelity and cost when supported.',
-    customWidthLabel: 'Width',
-    customHeightLabel: 'Height',
-    maskUrlLabel: 'Mask URL',
-    maskUrlPlaceholder: 'Optional public mask image URL',
-    enableWebSearchLabel: 'Web search',
-    enableWebSearchHint: 'Ground the request with current web context when supported.',
-    thinkingLevelLabel: 'Thinking level',
-    thinkingLevelHint: 'Increase reasoning depth when the engine supports it.',
-    limitGenerationsLabel: 'Limit generations',
-    limitGenerationsHint: 'Reduce extra exploratory generations when supported.',
-    toggleEnabled: 'Enabled',
-    toggleDisabled: 'Disabled',
-    estimatedCost: 'Estimated cost: {amount}',
-    referenceLabel: 'Reference images',
-    referenceHelper: 'Optional · up to {count} images',
-    referenceNote:
-      'Drag & drop, paste, upload from your device, or pull from your Library. These references are required for Edit mode.',
-    referenceButton: 'Library',
-    characterButton: 'Characters',
-    characterNote: 'Pick generated characters to use as identity anchors.',
-    characterHiddenNotice: 'Selected characters are saved and will be reused when you switch back to a compatible engine.',
-    characterLimitNotice: 'Only the first {count} character reference{suffix} will be used with this engine.',
-    referenceExpand: 'Show {count} more slots',
-    referenceCollapse: 'Hide extra slots',
-    referenceSlotLabel: 'Slot {index}',
-    referenceSlotHint: 'Drop image, click to upload, paste, or choose from your library.',
-    referenceSlotNameFallback: 'Image',
-    referenceSlotActions: {
-      upload: 'Upload',
-      library: 'Library',
-      replace: 'Replace',
-      remove: 'Remove',
-    },
-  },
-  history: {
-    eyebrow: 'Recent run',
-    empty: 'Launch a generation to populate your history. Each image variant appears below.',
-    runsLabel: {
-      zero: 'No runs yet',
-      other: '{count} runs',
-    },
-    noPreview: 'No preview',
-    loadMore: 'Load more',
-    refreshing: 'Refreshing history…',
-  },
-  runButton: {
-    idle: 'Generate images',
-    running: 'Generating…',
-  },
-  errors: {
-    onlyImages: 'Only image files are supported.',
-    unsupportedFormat: 'This engine accepts {formats}.',
-    uploadFailed: 'Unable to upload the selected image. Try again.',
-    fileTooLarge: 'Image exceeds {maxMB} MB. Compress it or choose a smaller file.',
-    unauthorized: 'Session expired — please sign in again and retry the upload.',
-    promptMissing: 'Prompt is required.',
-    referenceMissing: 'Provide at least one source image for edit mode.',
-    generic: 'Image generation failed.',
-  },
-  messages: {
-    success: 'Generated {count} image{suffix}.',
-    generatingInProgress: 'Generating in progress ({count})…',
-    savedToLibrary: 'Saved to Library.',
-    removedFromLibrary: 'Removed from Library.',
-  },
-  general: {
-    uploading: 'Uploading…',
-    cancelUpload: 'Cancel upload',
-    emptyEngines: 'No image engines available.',
-  },
-  library: {
-    button: 'Library',
-    empty: 'No assets saved yet. Upload images from the composer or the Library page.',
-    overlay: 'Use this image',
-    unsupported: 'Unsupported format',
-    supportedFormats: 'Supported formats: {formats}',
-    assetFallback: 'Asset',
-    tabs: {
-      all: 'All images',
-      upload: 'Uploaded images',
-      generated: 'Generated images',
-      character: 'Character assets',
-      angle: 'Angle assets',
-      upscale: 'Upscale assets',
-    },
-    modal: {
-      title: 'Select from library',
-      description: 'Choose an image you previously imported.',
-      close: 'Close',
-      error: 'Unable to load assets. Please retry.',
-      empty: 'No assets saved yet. Upload images from the composer or the Library page.',
-      emptyCompatible: 'No compatible assets for this model yet.',
-      emptyUploads: 'No uploaded images yet. Upload images from the composer or the Library page.',
-      emptyGenerated: 'No generated images saved yet. Save a generated image to see it here.',
-      emptyCharacter: 'No character assets saved yet. Generate one in Character Builder first.',
-      emptyAngle: 'No angle assets saved yet. Generate one in the Angle tool first.',
-      emptyUpscale: 'No upscale assets saved yet. Save an upscale result first.',
-    },
-  },
-  characterPicker: {
-    title: 'Select characters',
-    description: 'Choose from your Character Builder outputs. Selected characters are added as identity references.',
-    selected: 'Selected',
-    select: 'Select',
-    done: 'Done',
-    empty: 'No character references yet. Generate one in the Character Builder first.',
-    limitLabel: 'Up to {count} character reference{suffix}',
-  },
-  authGate: {
-    title: 'Create an account to render',
-    body: 'You can explore the image workspace, but starting a real image generation requires an account.',
-    primary: 'Create account',
-    secondary: 'Sign in',
-    close: 'Maybe later',
-  },
-};
-
-function isPlainRecord(value: unknown): value is Record<string, unknown> {
-  return Boolean(value) && typeof value === 'object' && !Array.isArray(value);
-}
-
-function mergeCopy<T>(defaults: T, overrides?: Partial<T> | null): T {
-  if (!isPlainRecord(defaults) || !isPlainRecord(overrides)) return defaults;
-  const next: Record<string, unknown> = { ...defaults };
-  Object.entries(overrides).forEach(([key, overrideValue]) => {
-    const defaultValue = next[key];
-    if (isPlainRecord(defaultValue) && isPlainRecord(overrideValue)) {
-      next[key] = mergeCopy(defaultValue, overrideValue);
-      return;
-    }
-    if (overrideValue !== undefined) {
-      next[key] = overrideValue;
-    }
-  });
-  return next as T;
-}
-
-function formatTemplate(template: string, values: Record<string, string | number>): string {
-  return Object.entries(values).reduce((result, [key, value]) => {
-    return result.replaceAll(`{${key}}`, String(value));
-  }, template);
-}
-
-function formatImageSizeLabel(value: string): string {
-  const normalized = value.trim().toLowerCase();
-  const labels: Record<string, string> = {
-    auto: 'Auto',
-    square: 'Square',
-    square_hd: 'Square HD',
-    portrait_4_3: 'Portrait 4:3',
-    portrait_16_9: 'Portrait 16:9',
-    landscape_4_3: 'Landscape 4:3',
-    landscape_16_9: 'Landscape 16:9',
-    custom: 'Custom size',
-  };
-  const parsed = parseGptImage2SizeKey(normalized);
-  if (parsed) {
-    return `${parsed.width} x ${parsed.height}`;
-  }
-  return labels[normalized] ?? value.toUpperCase();
-}
-
-function formatQualityLabel(value: string): string {
-  const normalized = value.trim().toLowerCase();
-  if (normalized === 'low') return 'Low';
-  if (normalized === 'medium') return 'Medium';
-  if (normalized === 'high') return 'High';
-  return value;
-}
-
-function buildCustomImageSize(width: string, height: string): GptImage2ImageSize | null {
-  const parsedWidth = Number(width);
-  const parsedHeight = Number(height);
-  if (!Number.isFinite(parsedWidth) || !Number.isFinite(parsedHeight)) {
-    return null;
-  }
-  return {
-    width: Math.round(parsedWidth),
-    height: Math.round(parsedHeight),
-  };
-}
-
-const MAX_REFERENCE_SLOTS = MAX_REFERENCE_IMAGES;
-const DEFAULT_VISIBLE_REFERENCE_SLOTS = 4;
-const QUICK_IMAGE_COUNT_OPTIONS = [1, 2, 4, 6, 8] as const;
-const DESKTOP_RAIL_MIN_WIDTH = 1088;
-const DEFAULT_UPLOAD_LIMIT_MB = Number.isFinite(Number(process.env.NEXT_PUBLIC_ASSET_MAX_IMAGE_MB ?? '25'))
-  ? Number(process.env.NEXT_PUBLIC_ASSET_MAX_IMAGE_MB ?? '25')
-  : 25;
-
-const IMAGE_COMPOSER_STORAGE_KEY = 'maxvideoai.image.composer.v1';
-const IMAGE_COMPOSER_STORAGE_VERSION = 4;
-const IMAGE_COMPOSER_STORAGE_DEBOUNCE_MS = 1200;
-
-function normalizeEngineToken(value: string): string {
-  return value.toLowerCase().replace(/[^a-z0-9]+/g, '');
-}
-
-function findImageEngine(engines: ImageEngineOption[], engineId: string): ImageEngineOption | null {
-  const trimmed = engineId.trim();
-  if (!trimmed) return null;
-  const direct = engines.find((entry) => entry.id === trimmed);
-  if (direct) return direct;
-  const token = normalizeEngineToken(trimmed);
-  if (!token) return null;
-  return (
-    engines.find((entry) => normalizeEngineToken(entry.id) === token) ??
-    engines.find((entry) => normalizeEngineToken(entry.name) === token) ??
-    engines.find((entry) => (entry.aliases ?? []).some((alias) => normalizeEngineToken(alias) === token)) ??
-    null
-  );
-}
-
-type PromptPreset = {
-  title: string;
-  prompt: string;
-  notes?: string;
-  mode: ImageGenerationMode;
-};
-
-export type ImageEngineOption = {
-  id: string;
-  name: string;
-  description?: string;
-  pricePerImage: number;
-  currency: string;
-  prompts: PromptPreset[];
-  modes: ImageGenerationMode[];
-  engineCaps: EngineCaps;
-  aliases: string[];
-};
-
-type HistoryEntry = {
-  id: string;
-  engineId: string;
-  engineLabel: string;
-  mode: ImageGenerationMode;
-  prompt: string;
-  createdAt: number;
-  description?: string | null;
-  images: GeneratedImage[];
-  jobId?: string | null;
-  aspectRatio?: string | null;
-};
-
-type UploadFailure = Error & { code?: string; maxMB?: number };
-
-type ReferenceSlotValue = {
-  id: string;
-  url: string;
-  previewUrl?: string;
-  name?: string;
-  status: 'ready' | 'uploading';
-  source?: 'upload' | 'library' | 'paste' | 'character';
-  width?: number | null;
-  height?: number | null;
-  characterReference?: CharacterReferenceSelection | null;
-};
-
-type PersistedReferenceSlot =
-  | {
-      url: string;
-      source?: ReferenceSlotValue['source'];
-      width?: number | null;
-      height?: number | null;
-      characterReference?: PersistedCharacterReference | null;
-    }
-  | null;
-type PersistedCharacterReference = CharacterReferenceSelection;
-
-type PersistedImageComposerState = {
-  version: number;
-  engineId: string;
-  mode: ImageGenerationMode;
-  prompt: string;
-  numImages: number;
-  aspectRatio: string | null;
-  resolution: string | null;
-  customImageSize: GptImage2ImageSize | null;
-  seed: number | null;
-  outputFormat: string | null;
-  quality: string | null;
-  maskUrl: string | null;
-  enableWebSearch: boolean;
-  thinkingLevel: string | null;
-  limitGenerations: boolean;
-  referenceSlots: PersistedReferenceSlot[];
-  characterReferences?: PersistedCharacterReference[];
-};
-
-function parsePersistedImageComposerState(value: string): PersistedImageComposerState | null {
-  try {
-    const raw = JSON.parse(value) as Partial<PersistedImageComposerState> | null;
-    if (!raw || typeof raw !== 'object') return null;
-    if (![1, 2, 3, IMAGE_COMPOSER_STORAGE_VERSION].includes(Number(raw.version))) return null;
-    if (typeof raw.engineId !== 'string' || raw.engineId.trim().length === 0) return null;
-    const mode = raw.mode === 't2i' || raw.mode === 'i2i' ? raw.mode : 't2i';
-    const prompt = typeof raw.prompt === 'string' ? raw.prompt : '';
-    const numImages =
-      typeof raw.numImages === 'number' && Number.isFinite(raw.numImages) ? Math.round(raw.numImages) : 1;
-    const aspectRatio = typeof raw.aspectRatio === 'string' ? raw.aspectRatio : null;
-    const resolution = typeof raw.resolution === 'string' ? raw.resolution : null;
-    const rawCustomImageSize =
-      raw.customImageSize && typeof raw.customImageSize === 'object'
-        ? (raw.customImageSize as Partial<GptImage2ImageSize>)
-        : null;
-    const customImageSize =
-      typeof rawCustomImageSize?.width === 'number' &&
-      Number.isFinite(rawCustomImageSize.width) &&
-      typeof rawCustomImageSize?.height === 'number' &&
-      Number.isFinite(rawCustomImageSize.height)
-        ? {
-            width: Math.round(rawCustomImageSize.width),
-            height: Math.round(rawCustomImageSize.height),
-          }
-        : null;
-    const seed =
-      typeof raw.seed === 'number' && Number.isFinite(raw.seed) ? Math.round(raw.seed) : null;
-    const outputFormat = typeof raw.outputFormat === 'string' ? raw.outputFormat : null;
-    const quality = typeof raw.quality === 'string' ? raw.quality : null;
-    const maskUrl = typeof raw.maskUrl === 'string' ? raw.maskUrl : null;
-    const enableWebSearch = raw.enableWebSearch === true;
-    const thinkingLevel = typeof raw.thinkingLevel === 'string' ? raw.thinkingLevel : null;
-    const limitGenerations = raw.limitGenerations === true;
-    const referenceSlotsRaw = Array.isArray(raw.referenceSlots) ? raw.referenceSlots : [];
-    const referenceSlots = referenceSlotsRaw
-      .slice(0, MAX_REFERENCE_SLOTS)
-      .map((entry): PersistedReferenceSlot => {
-        if (entry === null) return null;
-        if (!entry || typeof entry !== 'object') return null;
-        const record = entry as { url?: unknown; source?: unknown; characterReference?: unknown };
-        const url = typeof record.url === 'string' ? record.url.trim() : '';
-        if (!url || url.startsWith('blob:')) return null;
-        const source =
-          record.source === 'upload' ||
-          record.source === 'library' ||
-          record.source === 'paste' ||
-          record.source === 'character'
-            ? (record.source as ReferenceSlotValue['source'])
-            : undefined;
-        const width = typeof (record as { width?: unknown }).width === 'number' ? Math.round((record as { width: number }).width) : null;
-        const height = typeof (record as { height?: unknown }).height === 'number' ? Math.round((record as { height: number }).height) : null;
-        const characterReference = parseCharacterReferenceEntry(record.characterReference);
-        return { url, source, width, height, characterReference };
-      });
-    const characterReferencesRaw = Array.isArray(raw.characterReferences) ? raw.characterReferences : [];
-    const characterReferences = characterReferencesRaw.reduce<PersistedCharacterReference[]>((acc, entry) => {
-      const parsedEntry = parseCharacterReferenceEntry(entry);
-      if (!parsedEntry) return acc;
-      acc.push(parsedEntry);
-      return acc;
-    }, []);
-
-    return {
-      version: IMAGE_COMPOSER_STORAGE_VERSION,
-      engineId: raw.engineId.trim(),
-      mode,
-      prompt,
-      numImages,
-      aspectRatio,
-      resolution,
-      customImageSize,
-      seed,
-      outputFormat,
-      quality,
-      maskUrl,
-      enableWebSearch,
-      thinkingLevel,
-      limitGenerations,
-      referenceSlots,
-      characterReferences,
-    };
-  } catch {
-    return null;
-  }
-}
-
-type LibraryAsset = {
-  id: string;
-  url: string;
-  mime?: string | null;
-  width?: number | null;
-  height?: number | null;
-  size?: number | null;
-  source?: string | null;
-  createdAt?: string;
-};
-
-type AssetsResponse = {
-  ok: boolean;
-  assets: LibraryAsset[];
-};
-
-type ImageLibraryModalState = {
-  open: boolean;
-  slotIndex: number | null;
-  selectionMode: 'reference' | 'character';
-  initialSource: AssetLibrarySource;
-};
-
-type PricingEstimateResponse = {
-  ok: boolean;
-  pricing: PricingSnapshot;
-};
-
 interface ImageWorkspaceProps {
   engines: ImageEngineOption[];
-}
-
-function getCharacterReferenceLabel(reference: CharacterReferenceSelection): string {
-  if (reference.action === 'lighting-variant') return 'Lighting variant';
-  if (reference.action === 'full-body-fix') return 'Full-body fix';
-  if (reference.outputMode === 'character-sheet') return 'Character sheet';
-  return 'Portrait reference';
-}
-
-function formatCharacterReferenceDate(value?: string | null): string {
-  if (!value) return 'Recent';
-  const parsed = new Date(value);
-  if (Number.isNaN(parsed.getTime())) return 'Recent';
-  return parsed.toLocaleDateString(undefined, { month: 'short', day: 'numeric' });
-}
-
-function parseCharacterReferenceEntry(entry: unknown): PersistedCharacterReference | null {
-  if (!entry || typeof entry !== 'object') return null;
-  const record = entry as Partial<CharacterReferenceSelection>;
-  const id = typeof record.id === 'string' ? record.id.trim() : '';
-  const jobId = typeof record.jobId === 'string' ? record.jobId.trim() : '';
-  const imageUrl = typeof record.imageUrl === 'string' ? record.imageUrl.trim() : '';
-  if (!id || !jobId || !/^https?:\/\//i.test(imageUrl)) return null;
-  return {
-    id,
-    jobId,
-    imageUrl,
-    thumbUrl: typeof record.thumbUrl === 'string' ? record.thumbUrl : null,
-    prompt: typeof record.prompt === 'string' ? record.prompt : null,
-    createdAt: typeof record.createdAt === 'string' ? record.createdAt : null,
-    engineLabel: typeof record.engineLabel === 'string' ? record.engineLabel : null,
-    outputMode:
-      record.outputMode === 'portrait-reference' || record.outputMode === 'character-sheet'
-        ? record.outputMode
-        : null,
-    action:
-      record.action === 'generate' || record.action === 'full-body-fix' || record.action === 'lighting-variant'
-        ? record.action
-        : null,
-  };
-}
-
-function createCharacterReferenceSlot(
-  reference: CharacterReferenceSelection,
-  idPrefix = 'character'
-): ReferenceSlotValue {
-  return {
-    id: `${idPrefix}-${reference.id}`,
-    url: reference.imageUrl,
-    previewUrl: reference.thumbUrl ?? reference.imageUrl,
-    name: getCharacterReferenceLabel(reference),
-    status: 'ready',
-    source: 'character',
-    characterReference: reference,
-  };
-}
-
-function mergeCharacterReferencesIntoSlots(
-  slots: Array<ReferenceSlotValue | null>,
-  references: CharacterReferenceSelection[],
-  limit: number
-): Array<ReferenceSlotValue | null> {
-  if (!references.length || limit <= 0) return slots;
-  const next = slots.slice();
-  references.forEach((reference) => {
-    if (next.some((slot) => slot?.characterReference?.id === reference.id)) {
-      return;
-    }
-    const existingUrlIndex = next.findIndex((slot) => slot?.url === reference.imageUrl);
-    if (existingUrlIndex >= 0) {
-      next[existingUrlIndex] = {
-        ...(next[existingUrlIndex] as ReferenceSlotValue),
-        source: 'character',
-        characterReference: reference,
-        name: getCharacterReferenceLabel(reference),
-        previewUrl: next[existingUrlIndex]?.previewUrl ?? reference.thumbUrl ?? reference.imageUrl,
-      };
-      return;
-    }
-    const emptyIndex = next.slice(0, limit).findIndex((slot) => slot === null);
-    const targetIndex = emptyIndex >= 0 ? emptyIndex : -1;
-    if (targetIndex >= 0) {
-      next[targetIndex] = createCharacterReferenceSlot(reference);
-    }
-  });
-  return next;
-}
-
-function mapJobToHistoryEntry(job: Job): HistoryEntry | null {
-  const renderUrls = Array.isArray(job.renderIds)
-    ? job.renderIds.filter((url): url is string => typeof url === 'string' && url.length > 0)
-    : [];
-  const renderThumbUrls = Array.isArray(job.renderThumbUrls)
-    ? job.renderThumbUrls.filter((url): url is string => typeof url === 'string' && url.length > 0)
-    : [];
-  const heroOriginal = typeof job.heroRenderId === 'string' && job.heroRenderId.length ? job.heroRenderId : null;
-  const images: GeneratedImage[] =
-    renderUrls.length > 0
-      ? renderUrls.map((url, index) => ({
-          url,
-          thumbUrl: resolvePreferredMediaUrl(renderThumbUrls[index], index === 0 ? job.thumbUrl ?? null : null, url),
-        }))
-      : (() => {
-          const thumbUrl = resolvePreferredMediaUrl(job.thumbUrl, heroOriginal);
-          if (!thumbUrl || isPlaceholderMediaUrl(thumbUrl)) {
-            return [];
-          }
-          return [{ url: heroOriginal ?? thumbUrl, thumbUrl }];
-        })();
-  if (!images.length) return null;
-  const timestamp = Date.parse(job.createdAt ?? '');
-  return {
-    id: job.jobId,
-    jobId: job.jobId,
-    engineId: job.engineId ?? '',
-    engineLabel: job.engineLabel ?? job.engineId ?? 'Image generation',
-    mode: 't2i',
-    prompt: job.prompt ?? '',
-    createdAt: Number.isNaN(timestamp) ? Date.now() : timestamp,
-    description: job.message ?? null,
-    images,
-    aspectRatio: job.aspectRatio ?? null,
-  };
-}
-
-function buildPendingGroup({
-  id,
-  engineId,
-  engineLabel,
-  prompt,
-  count,
-  createdAt,
-}: {
-  id: string;
-  engineId: string;
-  engineLabel: string;
-  prompt: string;
-  count: number;
-  createdAt: number;
-}): GroupSummary {
-  const normalizedCount = Math.max(1, Math.round(count));
-  const previewCount = Math.min(4, normalizedCount);
-  const createdAtIso = new Date(createdAt).toISOString();
-  const members: GroupMemberSummary[] = Array.from({ length: previewCount }, (_, index) => ({
-    id: `${id}-pending-${index + 1}`,
-    engineId,
-    engineLabel,
-    durationSec: 0,
-    prompt,
-    createdAt: createdAtIso,
-    source: 'render',
-    status: 'pending',
-    iterationIndex: index,
-    iterationCount: normalizedCount,
-  }));
-
-  return {
-    id,
-    source: 'active',
-    splitMode: normalizedCount > 1 ? 'quad' : 'single',
-    count: normalizedCount,
-    totalPriceCents: null,
-    createdAt: createdAtIso,
-    hero: members[0],
-    previews: members.map((member) => ({
-      id: member.id,
-      thumbUrl: null,
-      videoUrl: null,
-      aspectRatio: null,
-    })),
-    members,
-  };
-}
-
-function buildCompletedGroup({
-  id,
-  engineId,
-  engineLabel,
-  prompt,
-  aspectRatio,
-  images,
-  createdAt,
-  totalPriceCents,
-  currency,
-}: {
-  id: string;
-  engineId: string;
-  engineLabel: string;
-  prompt: string;
-  aspectRatio: string | null;
-  images: GeneratedImage[];
-  createdAt: number;
-  totalPriceCents?: number | null;
-  currency?: string | null;
-}): GroupSummary {
-  const createdAtIso = new Date(createdAt).toISOString();
-  const members: GroupMemberSummary[] = images.map((image, index) => ({
-    id: `${id}-image-${index + 1}`,
-    jobId: id,
-    engineId,
-    engineLabel,
-    durationSec: 0,
-    priceCents: totalPriceCents ?? null,
-    currency: currency ?? null,
-    thumbUrl: image.url,
-    aspectRatio,
-    prompt,
-    status: 'completed',
-    progress: 100,
-    createdAt: createdAtIso,
-    source: 'render',
-  }));
-
-  return {
-    id,
-    source: 'active',
-    splitMode: members.length > 1 ? 'quad' : 'single',
-    count: members.length,
-    totalPriceCents: totalPriceCents ?? null,
-    currency: currency ?? null,
-    createdAt: createdAtIso,
-    hero: members[0],
-    previews: images.map((image, index) => ({
-      id: `${id}-preview-${index + 1}`,
-      thumbUrl: image.url,
-      videoUrl: null,
-      aspectRatio,
-      source: 'render',
-    })),
-    members,
-  };
 }
 
 export default function ImageWorkspace({ engines }: ImageWorkspaceProps) {
@@ -983,7 +161,7 @@ export default function ImageWorkspace({ engines }: ImageWorkspaceProps) {
   const [selectedPreviewEntryId, setSelectedPreviewEntryId] = useState<string | null>(null);
   const [selectedPreviewImageIndex, setSelectedPreviewImageIndex] = useState(0);
   const [pendingGroups, setPendingGroups] = useState<GroupSummary[]>([]);
-  const [isDesktopLayout, setIsDesktopLayout] = useState<boolean>(false);
+  const isDesktopLayout = useImageWorkspaceDesktopLayout();
   const [error, setError] = useState<string | null>(null);
   const [statusMessage, setStatusMessage] = useState<string | null>(null);
   const [areReferenceSlotsExpanded, setAreReferenceSlotsExpanded] = useState(false);
@@ -1004,21 +182,6 @@ export default function ImageWorkspace({ engines }: ImageWorkspaceProps) {
   ] = useState<[string, string, ImageGenerationMode, number, string, string, boolean, string, string, string] | null>(() =>
     engines[0] ? ['image-pricing', engines[0].id, 't2i', 1, '', '', false, '', '', ''] : null
   );
-
-  useEffect(() => {
-    if (typeof window === 'undefined') return undefined;
-    const mediaQuery = window.matchMedia(`(min-width: ${DESKTOP_RAIL_MIN_WIDTH}px)`);
-    const handleChange = (event: MediaQueryListEvent | MediaQueryList) => {
-      setIsDesktopLayout(event.matches);
-    };
-    handleChange(mediaQuery);
-    if (typeof mediaQuery.addEventListener === 'function') {
-      mediaQuery.addEventListener('change', handleChange);
-      return () => mediaQuery.removeEventListener('change', handleChange);
-    }
-    mediaQuery.addListener(handleChange);
-    return () => mediaQuery.removeListener(handleChange);
-  }, []);
 
   const engineCapsList = useMemo(() => engines.map((engine) => engine.engineCaps), [engines]);
   const selectedEngine = useMemo(
@@ -3263,372 +2426,5 @@ export default function ImageWorkspace({ engines }: ImageWorkspaceProps) {
       />
     ) : null}
     </>
-  );
-}
-
-function ImageLibraryModal({
-  open,
-  onClose,
-  onSelect,
-  onToggleCharacter,
-  selectedCharacterReferences,
-  characterSelectionLimit,
-  copy,
-  characterCopy,
-  selectionMode,
-  initialSource,
-  supportedFormats,
-  supportedFormatsLabel,
-  toolsEnabled,
-}: {
-  open: boolean;
-  onClose: () => void;
-  onSelect: (asset: LibraryAsset) => void;
-  onToggleCharacter: (reference: CharacterReferenceSelection) => void;
-  selectedCharacterReferences: CharacterReferenceSelection[];
-  characterSelectionLimit: number;
-  copy: ImageWorkspaceCopy['library'];
-  characterCopy: ImageWorkspaceCopy['characterPicker'];
-  selectionMode: 'reference' | 'character';
-  initialSource: AssetLibrarySource;
-  supportedFormats: string[];
-  supportedFormatsLabel: string;
-  toolsEnabled: boolean;
-}) {
-  const { t, locale } = useI18n();
-  const uiLocale = normalizeUiLocale(locale);
-  const [activeSource, setActiveSource] = useState<AssetLibrarySource>(initialSource);
-  const [isImporting, setIsImporting] = useState(false);
-  const [importError, setImportError] = useState<string | null>(null);
-  const importInputRef = useRef<HTMLInputElement | null>(null);
-  const isCharacterMode = selectionMode === 'character';
-  const importLabel = t(
-    'workspace.generate.assetLibrary.import',
-    uiLocale === 'fr' ? 'Importer' : uiLocale === 'es' ? 'Importar' : 'Import'
-  ) as string;
-  const importingLabel = t(
-    'workspace.generate.assetLibrary.importing',
-    uiLocale === 'fr' ? 'Import en cours…' : uiLocale === 'es' ? 'Importando…' : 'Importing…'
-  ) as string;
-  const importFailedLabel = t(
-    'workspace.generate.assetLibrary.importFailed',
-    uiLocale === 'fr' ? 'Import impossible. Réessayez.' : uiLocale === 'es' ? 'La importación falló. Inténtalo de nuevo.' : 'Import failed. Please try again.'
-  ) as string;
-  const swrKey = !open
-    ? null
-    : isCharacterMode
-      ? '/api/character-references?limit=60'
-      : activeSource === 'all'
-        ? '/api/user-assets?limit=60'
-        : `/api/user-assets?limit=60&source=${encodeURIComponent(activeSource)}`;
-  const { data, error, isLoading, mutate } = useSWR(swrKey, async (url: string) => {
-    const response = await authFetch(url);
-    if (isCharacterMode) {
-      const payload = (await response.json().catch(() => null)) as CharacterReferencesResponse | null;
-      if (!response.ok || !payload?.ok) {
-        let message: string | undefined;
-        if (payload && typeof payload.error === 'string') {
-          message = payload.error;
-        }
-        throw new Error(message ?? 'Failed to load character references');
-      }
-      return payload.characters;
-    }
-    const payload = (await response.json().catch(() => null)) as AssetsResponse | null;
-    if (!response.ok || !payload?.ok) {
-      let message: string | undefined;
-      if (payload && typeof payload === 'object' && 'error' in payload) {
-        const maybeError = (payload as { error?: unknown }).error;
-        if (typeof maybeError === 'string') {
-          message = maybeError;
-        }
-      }
-      throw new Error(message ?? 'Failed to load library');
-    }
-    return payload.assets;
-  });
-
-  useEffect(() => {
-    if (!open) return;
-    setActiveSource(initialSource);
-  }, [initialSource, open]);
-
-  const handleImportChange = useCallback(
-    async (event: ChangeEvent<HTMLInputElement>) => {
-      const file = event.currentTarget.files?.[0] ?? null;
-      event.currentTarget.value = '';
-      if (!file || isCharacterMode) return;
-
-      if (!file.type.startsWith('image/')) {
-        setImportError(t('workspace.image.errors.onlyImages', DEFAULT_COPY.errors.onlyImages) as string);
-        return;
-      }
-
-      setImportError(null);
-      setIsImporting(true);
-      try {
-        const preparedFile = await prepareImageFileForUpload(file, {
-          maxBytes: DEFAULT_UPLOAD_LIMIT_MB * 1024 * 1024,
-        });
-        const formData = new FormData();
-        formData.append('file', preparedFile, preparedFile.name);
-        const response = await authFetch('/api/uploads/image', {
-          method: 'POST',
-          body: formData,
-        });
-        const payload = await response.json().catch(() => null);
-        if (!response.ok || !payload?.ok || !payload?.asset?.url) {
-          const message = translateError({
-            code: typeof payload?.error === 'string' ? payload.error : null,
-            status: response.status,
-            message: importFailedLabel,
-          }).message;
-          throw new Error(message);
-        }
-
-        const uploadedAsset = payload.asset as LibraryAsset;
-        void mutate();
-        onSelect({
-          id: uploadedAsset.id ?? `library_${Date.now().toString(36)}`,
-          url: uploadedAsset.url,
-          mime: uploadedAsset.mime ?? null,
-          width: uploadedAsset.width ?? null,
-          height: uploadedAsset.height ?? null,
-          size: uploadedAsset.size ?? null,
-          source: uploadedAsset.source ?? 'upload',
-          createdAt: uploadedAsset.createdAt,
-        });
-      } catch (uploadError) {
-        setImportError(uploadError instanceof Error ? uploadError.message : importFailedLabel);
-      } finally {
-        setIsImporting(false);
-      }
-    },
-    [importFailedLabel, isCharacterMode, mutate, onSelect, t]
-  );
-
-  const assets = useMemo(() => {
-    if (isCharacterMode) return [];
-    return (((data ?? []) as LibraryAsset[])).filter((asset) =>
-      toolsEnabled ? true : asset.source !== 'character' && asset.source !== 'angle'
-    );
-  }, [data, isCharacterMode, toolsEnabled]);
-  const characters = useMemo(
-    () => (isCharacterMode ? (((data ?? []) as CharacterReferenceSelection[])) : []),
-    [data, isCharacterMode]
-  );
-  const availableSources = useMemo(
-    () =>
-      isCharacterMode
-        ? (['character'] as const satisfies readonly AssetLibrarySource[])
-        : toolsEnabled
-        ? (['all', 'upload', 'generated', 'character', 'angle', 'upscale'] as const satisfies readonly AssetLibrarySource[])
-        : (['all', 'upload', 'generated'] as const satisfies readonly AssetLibrarySource[]),
-    [isCharacterMode, toolsEnabled]
-  );
-
-  useEffect(() => {
-    if (!availableSources.some((source) => source === activeSource)) {
-      setActiveSource(initialSource);
-    }
-  }, [activeSource, availableSources, initialSource]);
-  const compatibilityByAssetId = useMemo(() => {
-    if (isCharacterMode) return new Map<string, boolean>();
-    const entries = assets.map((asset) => {
-      if (!supportedFormats.length) {
-        return [asset.id, true] as const;
-      }
-      const supportedByMime = isSupportedImageMime(supportedFormats, asset.mime);
-      if (supportedByMime != null) {
-        return [asset.id, supportedByMime] as const;
-      }
-      const inferredFormat = inferImageFormatFromUrl(asset.url);
-      return [asset.id, inferredFormat ? isSupportedImageFormat(supportedFormats, inferredFormat) : true] as const;
-    });
-    return new Map(entries);
-  }, [assets, isCharacterMode, supportedFormats]);
-
-  const compatibleAssets = useMemo(
-    () => (isCharacterMode ? [] : assets.filter((asset) => compatibilityByAssetId.get(asset.id) !== false)),
-    [assets, compatibilityByAssetId, isCharacterMode]
-  );
-  const browserAssets = useMemo(
-    () => {
-      if (isCharacterMode) {
-        return characters.map((character) => ({
-          id: character.id,
-          url: character.thumbUrl ?? character.imageUrl,
-          kind: 'image' as const,
-          source: 'character',
-          createdAt: character.createdAt ?? undefined,
-        }));
-      }
-      return compatibleAssets.map((asset) => ({
-        ...asset,
-        kind: 'image' as const,
-      }));
-    },
-    [characters, compatibleAssets, isCharacterMode]
-  );
-  const characterMap = useMemo(() => new Map(characters.map((character) => [character.id, character])), [characters]);
-  const selectedCharacterIds = useMemo(
-    () => new Set(selectedCharacterReferences.map((reference) => reference.id)),
-    [selectedCharacterReferences]
-  );
-  const emptyLabel =
-    isCharacterMode
-      ? characterCopy.empty
-      : compatibleAssets.length === 0 && assets.length > 0
-      ? copy.modal.emptyCompatible
-      : activeSource === 'generated'
-        ? copy.modal.emptyGenerated
-        : activeSource === 'upload'
-          ? copy.modal.emptyUploads
-          : activeSource === 'character'
-            ? copy.modal.emptyCharacter
-            : activeSource === 'angle'
-              ? copy.modal.emptyAngle
-              : activeSource === 'upscale'
-                ? copy.modal.emptyUpscale
-                : copy.modal.empty;
-  const supportedFormatsHint = isCharacterMode
-    ? formatTemplate(characterCopy.limitLabel, {
-        count: characterSelectionLimit,
-        suffix: characterSelectionLimit === 1 ? '' : 's',
-      })
-    : supportedFormats.length && supportedFormatsLabel.length
-      ? formatTemplate(copy.supportedFormats, { formats: supportedFormatsLabel })
-      : null;
-  const toolLinks = toolsEnabled
-    ? [
-        { href: '/app/image', label: 'Create image' },
-        { href: '/app/tools/angle', label: 'Change angle' },
-        { href: '/app/tools/character-builder', label: 'Character builder' },
-      ]
-    : [{ href: '/app/image', label: 'Create image' }];
-  const searchPlaceholder = t('workspace.library.browser.searchPlaceholder', 'Search assets…') as string;
-  const sourcesTitle = t('workspace.library.browser.sourcesTitle', 'Library') as string;
-  const toolsTitle = t('workspace.library.browser.toolsTitle', 'Create or transform') as string;
-  const toolsDescription = t(
-    'workspace.library.browser.toolsDescription',
-    'Open another workspace to prepare a better source before importing it here.'
-  ) as string;
-
-  const handleBackdropClick = (event: ReactMouseEvent<HTMLDivElement>) => {
-    if (event.target === event.currentTarget) {
-      onClose();
-    }
-  };
-
-  if (!open) {
-    return null;
-  }
-
-  return (
-    <div
-      className="fixed inset-0 z-[10050] flex items-center justify-center bg-surface-on-media-dark-60 px-3 py-6 sm:px-6"
-      role="dialog"
-      aria-modal="true"
-      onMouseDown={handleBackdropClick}
-    >
-      <div className="h-[92svh] w-full max-w-6xl overflow-hidden sm:h-[84vh]">
-        {!isCharacterMode ? (
-          <input
-            ref={importInputRef}
-            type="file"
-            accept="image/*"
-            className="sr-only"
-            onChange={handleImportChange}
-          />
-        ) : null}
-        <AssetLibraryBrowser
-          assetType="image"
-          layout="modal"
-          title={isCharacterMode ? characterCopy.title : copy.modal.title}
-          subtitle={supportedFormatsHint ?? (isCharacterMode ? characterCopy.description : copy.modal.description)}
-          onClose={onClose}
-          closeLabel={copy.modal.close}
-          assets={browserAssets}
-          isLoading={isLoading}
-          error={
-            importError ??
-            (error ? (isCharacterMode ? (error instanceof Error ? error.message : characterCopy.empty) : copy.modal.error) : null)
-          }
-          source={activeSource}
-          availableSources={availableSources}
-          sourceLabels={copy.tabs}
-          onSourceChange={setActiveSource}
-          headerActions={
-            !isCharacterMode ? (
-              <Button
-                type="button"
-                variant="outline"
-                size="sm"
-                className="rounded-full border-border bg-surface-2 px-3 text-sm text-text-secondary hover:bg-surface-3 hover:text-text-primary"
-                disabled={isImporting}
-                onClick={() => importInputRef.current?.click()}
-              >
-                {isImporting ? importingLabel : importLabel}
-              </Button>
-            ) : undefined
-          }
-          searchPlaceholder={searchPlaceholder}
-          sourcesTitle={sourcesTitle}
-          toolsTitle={toolsEnabled ? toolsTitle : undefined}
-          toolsDescription={toolsEnabled ? toolsDescription : undefined}
-          toolLinks={toolLinks}
-          emptyLabel={emptyLabel}
-          emptySearchLabel={copy.modal.empty}
-          renderAssetActions={(asset) => (
-            isCharacterMode ? (
-              (() => {
-                const character = characterMap.get(asset.id);
-                if (!character) return null;
-                const isSelected = selectedCharacterIds.has(character.id);
-                const isDisabled = !isSelected && selectedCharacterReferences.length >= characterSelectionLimit;
-                return (
-                  <Button
-                    type="button"
-                    size="sm"
-                    variant={isSelected ? 'primary' : 'outline'}
-                    disabled={isDisabled}
-                    onClick={() => onToggleCharacter(character)}
-                    className="min-h-0 h-9 rounded-full px-3 text-[11px] font-semibold"
-                  >
-                    {isSelected ? characterCopy.selected : characterCopy.select}
-                  </Button>
-                );
-              })()
-            ) : (
-              <Button
-                type="button"
-                size="sm"
-                variant="outline"
-                onClick={() => onSelect(asset as LibraryAsset)}
-                className="min-h-0 h-9 rounded-full px-3 text-[11px] font-semibold"
-              >
-                {copy.overlay}
-              </Button>
-            )
-          )}
-          renderAssetMeta={(asset) =>
-            isCharacterMode ? (
-              (() => {
-                const character = characterMap.get(asset.id);
-                if (!character) return null;
-                return (
-                  <>
-                    <span>{formatCharacterReferenceDate(character.createdAt)}</span>
-                    <span>{getCharacterReferenceLabel(character)}</span>
-                  </>
-                );
-              })()
-            ) : asset.createdAt ? (
-              <span>{new Date(asset.createdAt).toLocaleDateString()}</span>
-            ) : null
-          }
-        />
-      </div>
-    </div>
   );
 }
